Use parent.version instead of project.version
In a provider-specific release, the provider pom versions (i.e., project.version) will different from the core pom version
(unit-core/pom.xml's project.version). I don't know any direct way to access the project version of a peer module, but we can
use the parent.version in this case, because the root pom is supposed to follow the version number of the core module.

This only matters for releases that include provider specific tags. In all other cases, the project.version and parent.version are
the same value.